feat: moved reservation button to reservation list
