+++ libs/agency/feature/riders/src/lib/components/rider-profile/rider-profile.component.html
@@ -1,18 +1,6 @@
 <ng-container *ngIf="loadRiderDetails$ | async as rider; else loading">
   <section class="user-details">
     <nuverial-breadcrumb [breadCrumbs]="breadCrumbs"></nuverial-breadcrumb>
-    <nuverial-button
-      class="new-form-button"
-      ariaLabel="Schedule a ride"
-      [loading]="false"
-      (click)="createNewReservation()"
-      buttonStyle="outlined"
-      buttonType="submit"
-      colorTheme="primary"
-    >
-      <nuverial-icon class="black-icon" iconName="add" />
-      Schedule Ride
-    </nuverial-button>
     <dsg-rider-summary [rider]="rider"></dsg-rider-summary>
 
     <div class="detail-menu">
