+++ libs/agency/feature/riders/src/lib/components/rider-profile/rider-profile.component.ts
@@ -2,8 +2,7 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component, OnDestroy } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { ActivatedRoute, Router, RouterModule } from '@angular/router';
-import { WorkApiRoutesService } from '@dsg/shared/data-access/work-api';
+import { ActivatedRoute, RouterModule } from '@angular/router';
 import { EventsLogComponent } from '@dsg/shared/feature/events';
 import { FormRendererComponent } from '@dsg/shared/feature/form-nuv';
 import { RiderDetailsComponent, RiderProfileService, RiderSummaryComponent } from '@dsg/shared/feature/riders';
@@ -54,7 +53,6 @@ import { EMPTY, catchError, concatMap, of, switchMap, tap } from 'rxjs';
   templateUrl: './rider-profile.component.html',
 })
 export class RiderProfileComponent implements OnDestroy {
-  private readonly _reservationKey = 'MTAReservation';
   public breadCrumbs: INuverialBreadCrumb[] = [{ label: 'Back To Riders List', navigationPath: `/riders` }];
   public rider$ = this._riderProfileService.rider$.pipe(
     tap(rider => {
@@ -103,8 +101,6 @@ export class RiderProfileComponent implements OnDestroy {
     private readonly _riderProfileService: RiderProfileService,
     private readonly _route: ActivatedRoute,
     private readonly _nuverialSnackBarService: NuverialSnackBarService,
-    private readonly _workApiRoutesService: WorkApiRoutesService,
-    private readonly _router: Router,
   ) {}
 
   public ngOnDestroy(): void {
@@ -113,22 +109,5 @@ export class RiderProfileComponent implements OnDestroy {
 
   public trackByFn(index: number): number {
     return index;
-  }
-
-  public createNewReservation() {
-    const data = new Map();
-    data.set('riderId', this._riderProfileService.riderId);
-    data.set('riderUserId', this._riderProfileService.riderUserId);
-    this._workApiRoutesService
-      .createTransaction$(this._reservationKey, data)
-      .pipe(
-        tap(transaction => this._router.navigate([`${this.baseRoute}/transaction/${transaction.id}`])),
-        catchError(_error => {
-          this._nuverialSnackBarService.notifyApplicationError();
-
-          return EMPTY;
-        }),
-      )
-      .subscribe();
   }
 }
