+++ libs/shared/feature/riders/src/lib/components/locations/rider-saved-locations.component.html
@@ -1,5 +1,3 @@
-<!-- eslint-disable @angular-eslint/template/no-call-expression -->
-
 <div class="location-actions">
   <nuverial-button
     class="new-form-button"
