+++ libs/shared/feature/riders/src/lib/components/rides/rider-rides.component.scss
@@ -3,7 +3,10 @@
 :host {
   display: block;
 
-  .step-content {
-    margin: rem.fromPx(40px);
+  & > .reservation-actions {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+    padding: rem.fromPx(20px) rem.fromPx(30px);
   }
 }
