+++ libs/shared/feature/riders/src/lib/components/rides/rider-rides.component.ts
@@ -1,17 +1,20 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { TransactionModel } from '@dsg/shared/data-access/work-api';
+import { ActivatedRoute, Router } from '@angular/router';
+import { TransactionModel, WorkApiRoutesService } from '@dsg/shared/data-access/work-api';
 import { FormRendererComponent } from '@dsg/shared/feature/form-nuv';
 import {
   INuverialTab,
   NuverialBreadcrumbComponent,
+  NuverialButtonComponent,
   NuverialIconComponent,
+  NuverialSnackBarService,
   NuverialSpinnerComponent,
   NuverialTabKeyDirective,
   NuverialTabsComponent,
 } from '@dsg/shared/ui/nuverial';
 import { UntilDestroy } from '@ngneat/until-destroy';
-import { Observable } from 'rxjs';
+import { EMPTY, Observable, catchError, take, tap } from 'rxjs';
 import { RiderProfileService } from '../../services';
 import { RiderReservationListComponent } from '../rider-reservation-list';
 import { RiderRideHistoryComponent } from '../rider-ride-history';
@@ -29,6 +32,7 @@ import { RiderRideHistoryComponent } from '../rider-ride-history';
     NuverialTabKeyDirective,
     RiderReservationListComponent,
     RiderRideHistoryComponent,
+    NuverialButtonComponent,
   ],
   selector: 'dsg-rider-rides',
   standalone: true,
@@ -37,14 +41,50 @@ import { RiderRideHistoryComponent } from '../rider-ride-history';
 })
 export class RiderRidesComponent implements OnInit {
   public reservations$?: Observable<TransactionModel[]> = this._riderProfileService.reservations$;
+  public baseRoute = '';
 
   public tabs: INuverialTab[] = [
     { key: 'reservations', label: 'Reservations' },
     { key: 'history', label: 'Ride History' },
   ];
 
-  constructor(private readonly _riderProfileService: RiderProfileService) {}
+  private readonly _reservationKey = 'MTAReservation';
 
-  // eslint-disable-next-line @angular-eslint/no-empty-lifecycle-method, @typescript-eslint/no-empty-function
-  public ngOnInit() {}
+  constructor(
+    private readonly _riderProfileService: RiderProfileService,
+    private readonly _workApiRoutesService: WorkApiRoutesService,
+    private readonly _route: ActivatedRoute,
+    private readonly _router: Router,
+    private readonly _nuverialSnackBarService: NuverialSnackBarService,
+  ) {}
+
+  public createNewReservation() {
+    const data = new Map();
+    data.set('riderId', this._riderProfileService.riderId);
+    data.set('riderUserId', this._riderProfileService.riderUserId);
+    this._workApiRoutesService
+      .createTransaction$(this._reservationKey, data)
+      .pipe(
+        tap(transaction => this._router.navigate([`${this.baseRoute}/transaction/${transaction.id}`])),
+        catchError(_error => {
+          this._nuverialSnackBarService.notifyApplicationError();
+
+          return EMPTY;
+        }),
+      )
+      .subscribe();
+  }
+
+  public ngOnInit(): void {
+    this._route.paramMap
+      .pipe(
+        take(1),
+        tap(params => {
+          const riderId = params.get('recordId') ?? '';
+
+          this.baseRoute = `/riders/${riderId}`;
+        }),
+      )
+      .subscribe();
+  }
 }
